Fix json_compat test suite
The glob in `Rakefile` is `test_*.rb` so this suite wasn't ran.

Also minitest was complaining about `assert_equal nil, ...` so
I changed the generation a bit.
